feat(all): beatmap object, grade counter, friending, ranking
This update implements grade counting in the stats endpoint, as well as a new ranking endpoint, that allows beatmaps' approved status to be altered.

I've also made a check for if the beatmap isn't in the database or if the file isn't in the .osu directory, it'll use osu!api to fetch the things (maybe use a mirror instead??).

Other than that, small fixes and QOL changes here and there